+++ src/mem/vmem/vmem_device_memory/vmem_device_memory.c
@@ -5,16 +5,16 @@
  * @author: Anton Bondarev
  */
 
-#include <util/binalign.h>
-#include <util/log.h>
-
 #include <stddef.h>
+#include <stdint.h>
 #include <sys/mman.h>
 
 #include <hal/mmu.h>
 #include <kernel/task/resource/mmap.h>
 #include <mem/mmap.h>
 #include <mem/vmem.h>
+#include <util/binalign.h>
+#include <util/log.h>
 
 /**
  * @brief Map specific memory address to access device
@@ -27,14 +27,11 @@
  *
  * @return Pointer to allocated vmem or NULL if failed
  */
-void *mmap_device_memory(void *addr,
-		size_t len,
-		int prot,
-		int flags,
-		uint64_t physical){
+void *mmap_device_memory(void *addr, size_t len, int prot, int flags,
+    uintptr_t physical) {
 	/* TODO handle addr=NULL, region should be mapped to any addr
 	 * TODO handle flags anyhow */
-	uintptr_t start = (uintptr_t) (physical & ~MMU_PAGE_MASK);
+	uintptr_t start = (physical & ~(uintptr_t)MMU_PAGE_MASK);
 	uintptr_t end = binalign_bound(physical + len, MMU_PAGE_SIZE);
 
 	len = end - start;
@@ -50,11 +47,7 @@ void *mmap_device_memory(void *addr,
 		return NULL;
 	}
 
-	vmem_map_region(vmem_current_context(),
-				start,
-				start,
-				len,
-				prot);
+	vmem_map_region(vmem_current_context(), start, start, len, prot);
 	log_debug("phy (0x%" PRIxPTR ") mapped to %p", start, addr);
 
 	return addr;
